+++ ChangeLog.md
@@ -5,6 +5,7 @@
 * Fix encoding of JPEG2000 images with signed pixel data
 * Throttle queueing of PDUs to prevent out of memory errors for very large datasets
 * Better management of PDU buffer memory
+* Better handling for irregular specific character sets
 
 #### v1.0.22 (09/25/2012)
 * Add Offending Element tags to C-Move response output